[IMP] mail: Force email_from to a specific email address value instead of using sender email
Odoo is in general prepared to send e-mails using only the "author's" address as sender (From), and then use a "catchall" reply-to address in order to route back the answers (when someone replies to its notifications)

When you don't have manager rights on the email server (In this case, I only have the server with and email user address), what I could do is to force the email_from field to the email login user (that I was provided). From now on, every emails sending from odoo, no matter which user sends, they will be sent from only one email address.

This feature is done and mentioned in previous commit but the coding was still a mess and the commit is unclear. This commit finalized the code cleaning for this module, update the version and describe the feature more clearly
Changes to be committed:
	modified:   __init__.py		update removal of controllers folder
	modified:   __manifest__.py	update version name
	deleted:    controllers/__init__.py	remove unused folders
	deleted:    controllers/controllers.py	remove unused folders/files
	modified:   models/__init__.py	update removal of unused model
	deleted:    models/mail_message_monkeypatch.py	remove unused file